requirements.txt
comtypes==1.3.0
python-dotenv==1.0.1
h5py==3.11.0
# datetime==5.0
matplotlib==3.7.0
numpy==1.23.5
openpyxl==3.1.2
pandas==2.2.1
path==16.10.0
psutil==6.0.0
pyuff==2.4
scipy==1.10.0
seaborn==0.12.2
sympy==1.11.1